+++ api/src/models/referent.js
@@ -52,6 +52,26 @@ const Schema = new mongoose.Schema({
       description: "tentative de connexion. Max 15",
     },
   },
+  userIps: {
+    type: [String],
+    default: [],
+    documentation: {
+      description: "Liste des IP utilisées par l'utilisateur",
+    },
+  },
+  token2FA: {
+    type: String,
+    default: "",
+    documentation: {
+      description: "Token servant à la 2FA",
+    },
+  },
+  token2FAExpires: {
+    type: Date,
+    documentation: {
+      description: "Date limite de validité du token pour 2FA",
+    },
+  },
   acceptCGU: {
     type: String,
     enum: ["true", "false"],
@@ -208,24 +228,41 @@ const Schema = new mongoose.Schema({
     },
   },
 
+  deletedAt: { type: Date },
   createdAt: { type: Date, default: Date.now },
   updatedAt: { type: Date, default: Date.now },
 });
 
-Schema.pre("save", function (next) {
-  if (this.isModified("password") || this.isNew) {
-    bcrypt.hash(this.password, 10, (e, hash) => {
-      this.password = hash;
-      return next();
-    });
-  } else {
-    return next();
+Schema.pre("save", async function (next) {
+  if (this.isModified("password")) {
+    const hashedPassword = await bcrypt.hash(this.password, 10);
+    this.password = hashedPassword;
   }
+  return next();
+});
+
+Schema.pre("save", async function (next) {
+  if (this.isModified("userIps") || this.isNew) {
+    const _userIps = [];
+    for (let ip of this.userIps) {
+      const hashedIp = await bcrypt.hash(ip, 10);
+      _userIps.push(hashedIp);
+    }
+    this.userIps = _userIps;
+  }
+  return next();
 });
 
 Schema.methods.comparePassword = async function (p) {
   const user = await OBJ.findById(this._id).select("password");
   return bcrypt.compare(p, user.password || "");
+};
+
+Schema.methods.compareIps = async function (ip) {
+  const user = await OBJ.findById(this._id).select("userIps");
+  const promises = user.userIps.map((_ip) => bcrypt.compare(ip, _ip));
+  const responses = await Promise.all(promises);
+  return responses.some((e) => e);
 };
 
 //Sync with Sendinblue
@@ -273,6 +310,9 @@ Schema.plugin(patchHistory, {
     "/invitationExpires",
     "/loginAttempts",
     "/updatedAt",
+    "/userIps",
+    "/token2FA",
+    "/token2FAExpires",
   ],
 });
 
@@ -291,6 +331,9 @@ Schema.plugin(
       "loginAttempts",
       "updatedAt",
       "lastActivityAt",
+      "userIps",
+      "token2FA",
+      "token2FAExpires",
     ],
   }),
   MODELNAME,