+++ api/src/models/young.js
@@ -460,6 +460,26 @@ const Schema = new mongoose.Schema({
       description: "Mot de passe du volontaire",
     },
   },
+  userIps: {
+    type: [String],
+    default: [],
+    documentation: {
+      description: "Liste des IP utilisées par l'utilisateur",
+    },
+  },
+  token2FA: {
+    type: String,
+    default: "",
+    documentation: {
+      description: "Token servant à la 2FA",
+    },
+  },
+  token2FAExpires: {
+    type: Date,
+    documentation: {
+      description: "Date limite de validité du token pour 2FA",
+    },
+  },
   loginAttempts: {
     type: Number,
     default: 0,
@@ -1918,6 +1938,7 @@ const Schema = new mongoose.Schema({
   environmentInterest: { type: String },
   citizenshipInterest: { type: String },
 
+  deletedAt: { type: Date },
   createdAt: { type: Date, default: Date.now },
   updatedAt: { type: Date, default: Date.now },
 });
@@ -1933,9 +1954,29 @@ Schema.pre("save", function (next) {
   }
 });
 
+Schema.pre("save", async function (next) {
+  if (this.isModified("userIps") || this.isNew) {
+    if (!this.userIps) return next();
+    const _userIps = [];
+    for (let ip of this.userIps) {
+      const hashedIp = await bcrypt.hash(ip, 10);
+      _userIps.push(hashedIp);
+    }
+    this.userIps = _userIps;
+  }
+  return next();
+});
+
 Schema.methods.comparePassword = async function (p) {
   const user = await OBJ.findById(this._id).select("password");
   return bcrypt.compare(p, user.password || "");
+};
+
+Schema.methods.compareIps = async function (ip) {
+  const user = await OBJ.findById(this._id).select("userIps");
+  const promises = user.userIps.map((_ip) => bcrypt.compare(ip, _ip));
+  const responses = await Promise.all(promises);
+  return responses.some((e) => e);
 };
 
 //Sync with sendinblue
@@ -1989,6 +2030,9 @@ Schema.plugin(patchHistory, {
     "/statusPhase3UpdatedAt",
     "/statusPhase2ValidatedAt",
     "/statusPhase3ValidatedAt",
+    "/userIps",
+    "/token2FA",
+    "/token2FAExpires",
   ],
 });
 
@@ -2011,6 +2055,9 @@ Schema.plugin(
       "parent2Inscription2023Token",
       "updatedAt",
       "lastActivityAt",
+      "userIps",
+      "token2FA",
+      "token2FAExpires",
     ],
   }),
   MODELNAME,