+++ app/src/scenes/auth/desktop/signin2FA.jsx
@@ -1,0 +1,82 @@
+import React from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { toastr } from "react-redux-toastr";
+import RightArrow from "../../../assets/icons/RightArrow";
+import Input from "../../../components/inscription/input";
+import { setYoung } from "../../../redux/auth/actions";
+import api from "../../../services/api";
+import Error from "../../../components/error";
+import queryString from "query-string";
+import { useHistory } from "react-router-dom";
+import { BsShieldLock } from "react-icons/bs";
+
+export default function Signin() {
+  const [disabled, setDisabled] = React.useState(true);
+  const [loading, setLoading] = React.useState(false);
+  const [error, setError] = React.useState({});
+  const history = useHistory();
+  const [token2FA, setToken2FA] = React.useState("");
+
+  const dispatch = useDispatch();
+  const young = useSelector((state) => state.Auth.young);
+
+  const params = queryString.parse(location.search);
+  const { redirect, disconnected, email } = params;
+
+  React.useEffect(() => {
+    if (!young && disconnected === "1") toastr.error("Votre session a expiré", "Merci de vous reconnecter.", { timeOut: 10000 });
+    if (young) history.push("/" + (redirect || ""));
+  }, [young]);
+
+  const onSubmit = async ({ email, token }) => {
+    if (loading || disabled) return;
+    setLoading(true);
+    try {
+      setLoading(true);
+      const response = await api.post(`/young/signin-2fa`, { email, token_2fa: token });
+      setLoading(false);
+      if (response.token) api.setToken(response.token);
+      if (response.user) {
+        if (redirect?.startsWith("http")) return (window.location.href = redirect);
+        dispatch(setYoung(response.user));
+      }
+    } catch (e) {
+      setLoading(false);
+      toastr.error("Erreur détectée");
+    }
+  };
+
+  React.useEffect(() => {
+    if (token2FA) setDisabled(false);
+    else setDisabled(true);
+  }, [token2FA]);
+
+  return (
+    <div className="flex bg-[#F9F6F2] py-6">
+      <div className="mx-auto my-0 basis-[50%] bg-white px-[102px] py-[60px]">
+        {Object.keys(error).length > 0 && <Error {...error} onClose={() => setError({})} />}
+        <div className="mb-1 text-[32px] font-bold text-[#161616]">Me connecter</div>
+        <div className="mb-2 flex items-center gap-4">
+          <RightArrow />
+          <div className="flex items-center gap-2 text-blue-500 uppercase my-8 ">
+            <BsShieldLock className="text-blue-500 text-4xl" /> Authentification à deux facteurs
+          </div>
+        </div>
+        <div className="mb-1 flex flex-col gap-1 py-1">
+          <label className="text-base text-[#161616]">
+            Merci d'entrer le code transmis par e-mail à l'adresse <b>{email}</b> :
+          </label>
+          <Input placeholder="123abc" value={token2FA} onChange={(e) => setToken2FA(e)} />
+        </div>
+        <div className="flex w-full justify-end">
+          <button
+            disabled={disabled || loading}
+            className="flex cursor-pointer items-center justify-center bg-[#000091] px-3 py-2 text-white hover:border hover:border-[#000091] hover:bg-white hover:!text-[#000091]  disabled:cursor-default disabled:border-0 disabled:bg-[#E5E5E5] disabled:!text-[#929292]"
+            onClick={() => onSubmit({ email, token: token2FA })}>
+            Valider
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}