+++ app/src/scenes/auth/index.jsx
@@ -5,11 +5,13 @@ import MobileReset from "./mobile/reset";
 import MobileForgot from "./mobile/forgot";
 import MobileSignupInvite from "./mobile/signupInvite";
 import MobileSignin from "./mobile/signin";
+import MobileSignin2FA from "./mobile/signin2FA";
 
 import DesktopReset from "./desktop/reset";
 import DesktopForgot from "./desktop/forgot";
 import DesktopSignupInvite from "./desktop/signupInvite";
 import DesktopSignin from "./desktop/signin";
+import DesktopSignin2FA from "./desktop/signin2FA";
 
 import Connect from "./connect";
 import { SentryRoute } from "../../sentry";
@@ -28,6 +30,7 @@ const Render = ({ screen }) => {
     if (screen === "reset") return device === "desktop" ? <DesktopReset /> : <MobileReset />;
     if (screen === "forgot") return device === "desktop" ? <DesktopForgot /> : <MobileForgot />;
     if (screen === "auth") return device === "desktop" ? <DesktopSignin /> : <MobileSignin />;
+    if (screen === "2fa") return device === "desktop" ? <DesktopSignin2FA /> : <MobileSignin2FA />;
   }
 
   return (
@@ -50,6 +53,7 @@ export default function Index() {
       <SentryRoute path="/auth/reset" component={() => <Render screen="reset" />} />
       <SentryRoute path="/auth/forgot" component={() => <Render screen="forgot" />} />
       <SentryRoute path="/auth/connect" component={Connect} />
+      <SentryRoute path="/auth/2fa" component={() => <Render screen="2fa" />} />
       <SentryRoute path="/auth" component={() => <Render screen="auth" />} />
       <Redirect to={parentPath} /> {/* This will redirect to the parent path if no other Routes match */}
     </Switch>