+++ app/src/services/api.js
@@ -1,26 +1,12 @@
-import "isomorphic-fetch";
 import fetchRetry from "fetch-retry";
 import { apiURL } from "../config";
-import * as Sentry from "@sentry/react";
+import { capture } from "../sentry";
 
 let fetch = window.fetch;
-
-function jsonOrRedirectToSignIn(response) {
-  if (response.ok === false && response.status === 401) {
-    if (window?.location?.pathname !== "/auth") window.location.href = "/auth?disconnected=1";
-    // We need to return responses to prevent the promise from rejecting.
-    return { responses: [] };
-  }
-  return response.json();
-}
 
 class api {
   constructor() {
     this.token = "";
-  }
-
-  goToAuth() {
-    if (window?.location?.pathname !== "/auth") return (window.location.href = "/auth?disconnected=1");
   }
 
   getToken() {
@@ -46,6 +32,7 @@ class api {
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e);
         reject(e);
       }
     });
@@ -63,9 +50,16 @@ class api {
       headers: { "Content-Type": "application/x-ndjson", Authorization: `JWT ${this.token}` },
       body: [header, body].map((e) => `${JSON.stringify(e)}\n`).join(""),
     })
-      .then((r) => jsonOrRedirectToSignIn(r))
+      .then((response) => {
+        if (response.ok === false && response.status === 401) {
+          if (window?.location?.pathname !== "/auth") window.location.href = "/auth?disconnected=1";
+          // We need to return responses to prevent the promise from rejecting.
+          return { responses: [] };
+        }
+        return response.json();
+      })
       .catch((e) => {
-        Sentry.captureException(e);
+        capture(e, { extra: { body: body } });
         console.error(e);
         return { responses: [] };
       });
@@ -80,18 +74,24 @@ class api {
   }
 
   getAggregations(response) {
-    if (!response || !response.aggregations) return {};
-    const keys = Object.keys(response.aggregations);
-    if (!keys.length) return {};
+    try {
+      if (!response || !response.aggregations) return {};
+      const keys = Object.keys(response.aggregations);
+      if (!keys.length) return {};
 
-    if (response.aggregations[keys[0]].value !== undefined) return response.aggregations[keys[0]].value;
+      if (response.aggregations[keys[0]].value !== undefined) return response.aggregations[keys[0]].value;
 
-    let obj = {};
-    for (let i = 0; i < response.aggregations[keys[0]].buckets.length; i++) {
-      obj[response.aggregations[keys[0]].buckets[i].key] = response.aggregations[keys[0]].buckets[i].doc_count;
+      let obj = {};
+      for (let i = 0; i < response.aggregations[keys[0]].buckets.length; i++) {
+        obj[response.aggregations[keys[0]].buckets[i].key] = response.aggregations[keys[0]].buckets[i].doc_count;
+      }
+      return obj;
+    } catch (e) {
+      capture(e);
+      return;
     }
-    return obj;
   }
+
   openpdf(path, body) {
     return new Promise(async (resolve, reject) => {
       try {
@@ -106,13 +106,16 @@ class api {
           body: typeof body === "string" ? body : JSON.stringify(body),
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         if (response.status !== 200) return reject();
         const file = await response.blob();
         resolve(file);
       } catch (e) {
-        console.log(e);
+        capture(e, { extra: { body: body, path: path } });
         reject(e);
       }
     });
@@ -131,11 +134,15 @@ class api {
           headers: { "Content-Type": "application/json", Authorization: `JWT ${this.token}` },
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e, { extra: { path: path } });
         reject(e);
       }
     });
@@ -155,11 +162,15 @@ class api {
           body: typeof body === "string" ? body : JSON.stringify(body),
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e, { extra: { path: path, body: body } });
         reject(e);
       }
     });
@@ -187,11 +198,15 @@ class api {
           body: formData,
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e, { extra: { path: path, arr: arr } });
         reject(e);
       }
     });
@@ -218,11 +233,15 @@ class api {
           body: formData,
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e);
         reject(e);
       }
     });
@@ -241,11 +260,15 @@ class api {
           headers: { "Content-Type": "application/json", Authorization: `JWT ${this.token}` },
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         resolve(res);
       } catch (e) {
+        capture(e, { extra: { path: path } });
         reject(e);
       }
     });
@@ -265,7 +288,10 @@ class api {
           body: typeof body === "string" ? body : JSON.stringify(body),
         });
         if (response.status === 401) {
-          this.goToAuth();
+          if (window?.location?.pathname !== "/auth") {
+            window.location.href = "/auth?disconnected=1";
+            return;
+          }
         }
         const res = await response.json();
         if (response.status !== 200) {
@@ -273,6 +299,7 @@ class api {
         }
         resolve(res);
       } catch (e) {
+        capture(e, { extra: { path: path, body: body } });
         reject(e);
       }
     });